Changed all tile sizes from int to double

+++ src/authoring_environment/CircleTile.java
@@ -8,24 +8,24 @@ import javafx.scene.shape.Shape;
 
 public class CircleTile extends SuperTile{
 
-	public CircleTile(int size, Point2D loc) {
+	public CircleTile(double size, Point2D loc) {
 		super(size, loc);
 	}
 
 	@Override
-	protected Shape makeShape(int size) {
+	protected Shape makeShape(double size) {
 		return new Circle(size/2);
 	}
 
 	@Override
-	protected Point2D calculateCoord(int size, Point2D loc) {
+	protected Point2D calculateCoord(double size, Point2D loc) {
 		double xCenter=loc.getY()*size+size/2;
 		double yCenter=loc.getX()*size+size/2;
 		return new Point2D.Double(xCenter, yCenter);
 	}
 
 	@Override
-	protected Point2D calculateImageCoord(int size, Point2D loc) {
+	protected Point2D calculateImageCoord(double size, Point2D loc) {
 		Point2D layoutCenter=calculateCoord(size,loc);
 		double imageX=layoutCenter.getX()-size/Math.sqrt(2);
 		double imageY=layoutCenter.getY()-size/Math.sqrt(2);
@@ -33,7 +33,7 @@ public class CircleTile extends SuperTile{
 	}
 
 	@Override
-	protected int calculateImageSize(int size) {
+	protected double calculateImageSize(double size) {
 		return (int)Math.round(2*size/Math.sqrt(2));
 	}
