+++ src/authoring_environment/SuperTile.java
@@ -18,7 +18,7 @@ import javafx.scene.shape.Shape;
 public abstract class SuperTile extends Group {
 	private Shape myShape;
 //	private int mySize;
-	private int myImageSize;
+	private double myImageSize;
 	private Point2D myCoordinates;
 	private Point2D myLocation;
 	private Point2D myImageCoord;
@@ -33,7 +33,7 @@ public abstract class SuperTile extends Group {
 	 * @param size Tile size
 	 * @param loc Grid coordination
 	 */
-	public SuperTile(int size,Point2D loc) {
+	public SuperTile(double size,Point2D loc) {
 		
 		myLocation=loc;
 		
@@ -45,7 +45,7 @@ public abstract class SuperTile extends Group {
 	}
 	
 
-	protected void makeShapeTile(int size, Point2D loc){
+	protected void makeShapeTile(double size, Point2D loc){
 		myShape=makeShape(size);
 		makeHighlight(size);
 		setCheckeredColor((int)loc.getX(),(int)loc.getY(),myShape);
@@ -62,7 +62,7 @@ public abstract class SuperTile extends Group {
 	}
 
 
-	private void makeHighlight(int size) {
+	private void makeHighlight(double size) {
 		myHighlight=makeShape(size);
 		myHighlight.setFill(Color.web("#0000FF", 0.3));
 		myHighlight.setVisible(false);
@@ -89,13 +89,13 @@ public abstract class SuperTile extends Group {
 	}
 
 
-	protected abstract Shape makeShape(int size);
+	protected abstract Shape makeShape(double size);
 	
-	protected abstract Point2D calculateCoord(int size, Point2D loc);
+	protected abstract Point2D calculateCoord(double size, Point2D loc);
 	
-	protected abstract Point2D calculateImageCoord(int size, Point2D loc);
+	protected abstract Point2D calculateImageCoord(double size, Point2D loc);
 
-	protected abstract int calculateImageSize(int size);
+	protected abstract double calculateImageSize(double size);
 
 
 	private void setCheckeredColor(int row, int col, Shape shape) {
@@ -108,7 +108,7 @@ public abstract class SuperTile extends Group {
 		shape.setStroke(Color.GRAY);
 		shape.setStrokeWidth(0.75);
 	}
-	private ImageView initImageView(int size) {
+	private ImageView initImageView(double size) {
 		ImageView imgView = new ImageView();
 		imgView.setFitHeight(size);
 		imgView.setFitWidth(size);
